Feat/fix-service-and-controllers (#16)
* Maintenance

* Fixed the build

* Updated http methods
