+++ src/main/java/ro/unibuc/hello/service/ToDoService.java
@@ -6,11 +6,13 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import lombok.AllArgsConstructor;
+import ro.unibuc.hello.service.SharingService;
 import ro.unibuc.hello.data.*;
 import ro.unibuc.hello.dto.request.*;
 import ro.unibuc.hello.dto.request.ToDoListDto;
 import ro.unibuc.hello.dto.response.*;
 import ro.unibuc.hello.exception.EntityNotFoundException;
+
 
 @AllArgsConstructor
 @Service
@@ -35,6 +37,9 @@ public class ToDoService {
     private final RequestRepository requestRepository;
 
     private final ModelMapper modelMapper;
+
+    @Autowired
+    private final SharingService sharingService;
 
 
     public ItemResponseDto createItem(ItemDto itemDto) {
@@ -97,106 +102,6 @@ public class ToDoService {
         return true;
     }
 
-    public boolean createBind(String username, String toDoList, boolean isOwner) {
-
-       try{
-
-        Optional<UserListEntity> existingUserList = userListRepository.findByUsernameAndToDoListAndIsOwner(username, toDoList, isOwner);
-        
-        if (existingUserList.isPresent()) {
-
-            return false;
-        }
-
-        userListRepository.save(new UserListEntity(username,toDoList,isOwner));
-
-       }
-       catch (Exception exception)
-       {
-        return false;
-       }
-       return true;
-    }
-
-    public boolean deleteBind(String username, String toDoList) {
-        try {
-           
-            Optional<UserListEntity> existingUserList = userListRepository.findByUsernameAndToDoList(username, toDoList);
-    
-            if (existingUserList.isPresent()) {
-               
-                userListRepository.delete(existingUserList.get());
-                return true;
-            } else {
-                
-                return false;
-            }
-        } catch (Exception exception) {
-           
-            return false;
-        }
-    }
-    
-        
-    public RequestResponseDto createRequest(RequestDto requestDto) {
-        try {
-
-            var user = userService.getAuthenticatedUser();
-    
-            RequestEntity newRequest = new RequestEntity(user.getUsername(), requestDto.getToDoList(), requestDto.getDescription());
-            requestRepository.save(newRequest);
-            return modelMapper.map(newRequest, RequestResponseDto.class);  
-    
-        } catch (Exception exception) {
-            
-            throw new EntityNotFoundException("request");
-        }
-    }
-    
-
-    public RequestResponseDto acceptRequest(RequestDto requestDto) {
-        try {
-           
-            Optional<RequestEntity> requestOpt = requestRepository.findByUsernameAndToDoList(requestDto.getUsername(), requestDto.getToDoList());
-            
-            if (requestOpt.isPresent()) {
-                
-                createBind(requestDto.getUsername(), requestDto.getToDoList(), false);
-                
-                requestRepository.delete(requestOpt.get());
-                
-                return modelMapper.map(requestDto,RequestResponseDto.class);
-            } else {
-                throw new EntityNotFoundException("request"); 
-            }
-        } catch (Exception exception) {
-           
-            System.out.println("Error accepting request: " + exception.getMessage());
-            throw new EntityNotFoundException("request"); 
-        }
-    }
-
-
-    public boolean denyRequest(RequestDto requestDto) {
-        try {
-           
-            Optional<RequestEntity> requestOpt = requestRepository.findByUsernameAndToDoList(requestDto.getUsername(), requestDto.getToDoList());
-            
-            if (requestOpt.isPresent()) {
-                
-                requestRepository.delete(requestOpt.get());
-
-                return true;
-                
-            } else {
-                return false;  
-            }
-        } catch (Exception exception) {
-           
-            System.out.println("Error accepting request: " + exception.getMessage());
-            return false;
-        }
-    }
 
     public ToDoListResponseDto createToDoList(ToDoListDto toDoListDto) {
         try
@@ -204,7 +109,7 @@ public class ToDoService {
             var toDoList = modelMapper.map(toDoListDto, ToDoListEntity.class);
             toDoListRepository.save(toDoList);
             UserEntity user = userService.getAuthenticatedUser();
-            createBind(user.getUsername(), toDoList.getName(), true);
+            sharingService.createBind(user.getUsername(), toDoList.getName(), true);
             return modelMapper.map(toDoList, ToDoListResponseDto.class);
         }
         catch (Exception exception)
@@ -240,6 +145,7 @@ public class ToDoService {
 
         return true;
     }
+
     public ToDoListCollectionDto getMyToDoLists() {
         return new ToDoListCollectionDto();
     }
